split module_graph into directory
